consumidor/requirements.txt
paho-mqtt==1.6.1
mysql-connector-python==9.0.0

